Law & Policy Chatbot/requirements.txt
absl-py==2.1.0
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
arabic-reshaper==3.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
babel==2.16.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.12.3
billiard==4.2.1
bitsandbytes==0.42.0
bleach==6.1.0
build==1.3.0
cachetools==5.5.0
celery==5.4.0
certifi==2025.4.26
cffi==1.17.1
channels==4.1.0
chardet==5.2.0
charset-normalizer==3.4.1
chromadb==1.1.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coloredlogs==15.0.1
comm==0.2.2
constantly==23.10.4
contourpy==1.3.1
cryptography==43.0.1
cssselect2==0.7.0
cycler==0.12.1
Cython==0.29.37
daphne==4.1.2
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
decouple==0.0.7
defusedxml==0.7.1
distro==1.9.0
dj-database-url==2.3.0
dj-email-url==1.0.6
Django==4.2.17
django-appconf==1.0.6
django-bootstrap-datepicker-plus==5.0.5
django-cache-url==3.4.5
django-countries==7.6.1
django-environ==0.12.0
django-filter==24.3
django-formtools==2.5.1
django-js-asset==3.0.1
django-measurement==3.2.4
django-mptt==0.16.0
django-paypal==2.1
django-phonenumber-field==8.0.0
django-prices==2.3.0
django-select2==8.2.1
django-stubs-ext==5.1.1
django-summernote==0.8.20.0
django-tinymce==4.1.0
djangorestframework==3.15.2
drf-yasg==1.21.7
durationpy==0.10
enmerkar==0.7.1
executing==2.2.0
faiss-cpu==1.12.0
fastjsonschema==2.21.1
filelock==3.19.1
flatbuffers==24.12.23
fonttools==4.57.0
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.9.0
gast==0.6.0
google-ai-generativelanguage==0.6.9
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.1
google-pasta==0.2.0
googleapis-common-protos==1.65.0
graphql==0.0.4
graphql-core==2.3.2
greenlet==3.0.3
grpcio==1.69.0
grpcio-status==1.66.1
h11==0.14.0
h5py==3.12.1
hf-xet==1.1.10
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.1
huggingface-hub==0.35.3
humanfriendly==10.0
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
incremental==24.7.2
inflection==0.5.1
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.6
isoduration==20.11.0
jaeger-client==4.8.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.14
keras==3.11.3
kiwisolver==1.4.8
kombu==5.4.2
kubernetes==33.1.0
langchain==0.3.27
langchain-community==0.3.30
langchain-core==0.3.76
langchain-groq==0.3.8
langchain-huggingface==0.3.1
langchain-text-splitters==0.3.11
langsmith==0.4.31
libclang==18.1.1
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
markdown2==2.5.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
measurement==3.2.2
mistune==3.1.3
ml-dtypes==0.3.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
namex==0.0.8
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.4.0
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.3.1
ollama==0.3.3
onnxruntime==1.23.0
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentracing==2.4.0
opt_einsum==3.4.0
optree==0.13.1
orjson==3.11.3
oscrypto==1.3.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
paypal==1.2.5
paypal-checkout-serversdk==1.0.3
paypalhttp==1.0.1
paypalrestsdk==1.13.3
pexpect==4.9.0
pillow==11.1.0
pillow-avif-plugin==1.0.1
platformdirs==4.3.7
playwright==1.47.0
posthog==5.4.0
prices==1.1.1
prometheus_client==0.21.1
promise==2.3
prompt_toolkit==3.0.48
propcache==0.3.2
proto-plus==1.24.0
protobuf==5.29.5
psutil==7.0.0
psycopg==3.2.3
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybase64==1.4.2
pycparser==2.22
pydantic==2.9.1
pydantic-settings==2.11.0
pydantic_core==2.23.3
pyee==12.0.0
Pygments==2.19.1
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
pyOpenSSL==24.2.1
pyparsing==3.1.4
pypdf==4.3.1
PyPika==0.48.9
pypng==0.20220715.0
pyproject_hooks==1.2.0
PySocks==1.7.1
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-json-logger==3.3.0
pytimeparse==1.1.8
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.4.0
qrcode==7.4.2
redis==5.2.0
referencing==0.36.2
regex==2025.9.18
reportlab==4.2.2
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.24.0
rsa==4.9
Rx==1.6.3
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.14.1
seaborn==0.13.2
selenium==4.24.0
Send2Trash==1.8.3
sentence-transformers==5.1.1
sentencepiece==0.2.1
sentry-sdk==2.19.2
service-identity==24.1.0
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.43
sqlparse==0.5.1
stack-data==0.6.3
stripe==10.11.0
svglib==1.5.1
sympy==1.13.3
tenacity==9.1.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
termcolor==2.5.0
terminado==0.18.1
tf_keras==2.16.0
threadloop==1.0.2
threadpoolctl==3.5.0
thrift==0.21.0
tinycss2==1.2.1
tokenizers==0.22.1
torch==2.2.2
tornado==6.4.2
tqdm==4.66.5
traitlets==5.14.3
transformers==4.56.2
trio==0.26.2
trio-websocket==0.11.1
Twisted==24.7.0
txaio==23.1.1
typer==0.19.2
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uri-template==1.3.0
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.3.0
uvicorn==0.37.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.13
webcolors==24.11.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.14
wrapt==1.17.1
wsproto==1.2.0
xhtml2pdf==0.2.16
yarl==1.20.1
zipp==3.23.0
zope.interface==7.0.3
zstandard==0.25.0
